bots/binance/testnet/requirements.txt
# Binance Futures Testnet Trading Bot Requirements
# This file provides pip-compatible requirements for users who prefer pip over uv

# Core Nautilus Trading Framework
nautilus_trader>=1.192.0

# HTTP client for API calls
aiohttp>=3.8.0

# Configuration and environment
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0

# Async utilities
asyncio-throttle>=1.0.0

# Testing (development)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Logging and monitoring
structlog>=23.0.0

# Type checking (development)
mypy>=1.0.0

# Code formatting (development)
black>=23.0.0
isort>=5.12.0

# Documentation (development)
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Note: This project uses uv for dependency management by default.
# For uv installation and usage, see README.md
# To install with pip instead: pip install -r requirements.txt
